♻️  [#81] Refactor: types 디렉토리 생성 후 InputProps 파일 이동
