+++ apps/what-today/src/components/login/PasswordInput.tsx
@@ -1,7 +1,7 @@
 import { EyeIcon, EyeOffIcon, Input } from '@what-today/design-system';
 import { memo, useState } from 'react';
 
-import type InputProps from '@/schemas/InputProps';
+import type InputProps from '@/types/InputProps';
 
 function PasswordInput({ value, onChange }: InputProps) {
   const [isPasswordVisible, setIsPasswordVisible] = useState<boolean>(false);
