+++ apps/what-today/src/types/InputProps.ts
